Data.Deque

Description

A Deque is a spine-strict double-ended queue that incrementally rebalances itself when elements are added or removed.

Laziness

Evaluating the root of the queue (i.e. (_ :: Deque a)) to weak head normal form evaluates equal extents of the back and the front of the queue to normal form, while the remaining parts of both sides are represented as at most two thunks.

Functions do not perform any additional evaluations unless their documentation directly specifies so.

Performance

Each function's time complexity is provided in the documentation.

\(n\) refers to the total number of elements in the queue. Parts of the heap are denoted using subscripts: \(n_L\) refers to the left side, \(n_R\) to the right side, and \(n_M\) refers to the total number of elements collected with the Monoid.

Implementation

The implementation is specialized for \(c = 2\).

Description of the real-time double-ended queue can be found within the following paper:

Chris Okasaki, "Purely Functional Data Structures", September 1996, pages 57-59, https://www.cs.cmu.edu/~rwh/students/okasaki.pdf

Itself

data Deque a where

Spine-strict double-ended queue.

Bundled Patterns

pattern Empty :: Deque a
\(\mathcal{O}(1)\). Bidirectional pattern matching an empty queue.

pattern (:<|) :: a -> Deque a -> Deque a infixr 5
\(\mathcal{O}(1)\). Bidirectional pattern analyzing the front of a non-empty queue.

pattern (:|>) :: Deque a -> a -> Deque a infixl 5
\(\mathcal{O}(1)\). Bidirectional pattern analyzing the back of a non-empty queue.

Construct

empty :: Deque a

\(\mathcal{O}(1)\). Empty queue.

singleton :: a -> Deque a

\(\mathcal{O}(1)\). Queue with a single element.

fromListF :: [a] -> Deque a

\(\mathcal{O}(n)\). Construct a queue from a list. Element order matches list order, with the front of the queue corresponding to the head of the list.

fromListB :: [a] -> Deque a

\(\mathcal{O}(n)\). Construct a queue from a list. Element order matches list order, with the front of the queue corresponding to the tail of the list.

Front

Insert

(<|) :: a -> Deque a -> Deque a infixr 5

\(\mathcal{O}(1)\). Add an element to the front of the queue.

View

data ViewF a

View of the front of the queue.

Constructors

a :< !(Deque a)

viewF :: Deque a -> Maybe (ViewF a)

\(\mathcal{O}(1)\). Analyze the front of the queue.

Back

Insert

(|>) :: Deque a -> a -> Deque a infixl 5

\(\mathcal{O}(1)\). Add an element to the back of the queue.

View

data ViewB a

View of the back of the queue.

Constructors

!(Deque a) :> a

viewB :: Deque a -> Maybe (ViewB a)

\(\mathcal{O}(1)\). Analyze the back of the queue.

Full queue

Size

null :: Deque a -> Bool

\(\mathcal{O}(1)\). Check if the queue is empty.

size :: Deque a -> Int

\(\mathcal{O}(1)\). Calculate the number of elements in the queue. The returned number is guaranteed to be non-negative.

Reverse

reverse :: Deque a -> Deque a

\(\mathcal{O}(1)\). Reverse the element order.

Fold

Left-to-right

foldl :: (b -> a -> b) -> b -> Deque a -> b

\(\mathcal{O}(n_R)\). Fold the queue left-to-right.

foldl' :: (b -> a -> b) -> b -> Deque a -> b

\(\mathcal{O}(n)\). Fold the queue left-to-right with a strict accumulator.

Right-to-left

foldr :: (a -> b -> b) -> b -> Deque a -> b

\(\mathcal{O}(n_L)\). Fold the queue right-to-left.

foldr' :: (a -> b -> b) -> b -> Deque a -> b

\(\mathcal{O}(n)\). Fold the queue right-to-left with a strict accumulator.

Monoid

foldMap :: Monoid m => (a -> m) -> Deque a -> m

\(\mathcal{O}(n_M)\). Map each element of the queue into a monoid, and combine the results with (<>).

List

toList :: Deque a -> [a]

\(\mathcal{O}(n)\). Construct a list from a queue. Element order is arbitrary.

toListF :: Deque a -> [a]

\(\mathcal{O}(n)\). Construct a list from a queue. Element order matches queue order, with the head of the list corresponding to the front of the queue.

toListB :: Deque a -> [a]

\(\mathcal{O}(n)\). Construct a list from a queue. Element order matches queue order, with the head of the list corresponding to the back of the queue.