Data.Deque.Debug

Description

Safe functions for datatype introspection.

Show

showDeque :: (a -> ShowS) -> Deque a -> ShowS

\(\mathcal{O}(n)\). Shows the internal structure of the queue.

Eq

eqDeque :: (a -> b -> Bool) -> Deque a -> Deque b -> Bool

\(\mathcal{O}(n)\). Checks whether the internal representations of two queues are equal.

Validate

data Validity

Whether the queue is well-formed.

Constructors

Valid

Invalid Reason

data Reason

Reason for why the queue is considered malformed.

Constructors

Imbalanced
\(|F| \gt 2 |R| + 1\) or \(|R| \gt 2 |F| + 1\).

BadLength
Stored length does not match actual length on either side.

BadSchedule
Either schedule is too large in regards to stored lengths.

validate :: Deque a -> Validity

O(n). Checks whether the queue is well-formed.