Data.Queue.Unsafe

Description

Data structure internals, helper operations and unsafe functions.

Itself

data Queue a

Spine-strict queue.

Constructors

Queue
Invariants: \(|S| = |F| - |R|\).

Fields

![a]

\(F\). Front of the queue, in normal order.

!(List' a)

\(R\). Rear of the queue, in reverse order.

[a]

\(S\). Schedule: reference to some point in \(F\).

data List' a

Spine-strict list.

Constructors

Nil'

Cons' a !(List' a)

Invariant

data MalformedQueue

Exception thrown if the internal length invariant is broken, i.e. \(|F| \lt |R|\).

Constructors

MalformedQueue

Front

Insert

(<|) :: a -> Queue a -> Queue a infixr 5

\(\mathcal{O}(1)\). Add an element to the front of the queue.

Makes \(S\) diverge from \(F\) to preserve the invariant, doubling allocation overhead for this element.