Data.Steque

Description

A Steque is a spine-strict stack-ended queue that reverses its back in one go once the front runs dry.

Laziness

Evaluating the root of the queue (i.e. (_ :: Steque a)) to weak head normal form evaluates the entire spine of the queue to normal form.

Functions do not perform any additional evaluations unless their documentation directly specifies so.

Performance

Each function's time complexity is provided in the documentation.

\(n\) refers to the total number of elements in the queue. Parts of the heap are denoted using subscripts: \(n_L\) refers to the left side, \(n_R\) to the right side, and \(n_M\) refers to the total number of elements collected with the Monoid.

Implementation

Description of the amortized queue can be found within the following paper:

Chris Okasaki, "Purely Functional Data Structures", September 1996, pages 15-18, https://www.cs.cmu.edu/~rwh/students/okasaki.pdf

Itself

data Steque a where

Spine-strict stack-ended queue.

Bundled Patterns

pattern Empty :: Steque a
\(\mathcal{O}(1)\). Bidirectional pattern synonym matching an empty queue.

pattern (:<|) :: a -> Steque a -> Steque a infixr 5
\(\mathcal{O}(1)\) construction and amortized match, \(\mathcal{O}(n)\) worst-case match. Bidirectional pattern synonym analyzing the front of a non-empty queue.

Construct

empty :: Steque a

\(\mathcal{O}(1)\). Empty queue.

singleton :: a -> Steque a

\(\mathcal{O}(1)\). Queue with a single element.

fromListF :: [a] -> Steque a

\(\mathcal{O}(n)\). Construct a queue from a list. Element order matches list order, with the front of the queue corresponding to the head of the list.

Front

Insert

(<|) :: a -> Steque a -> Steque a infixr 5

\(\mathcal{O}(1)\). Add an element to the front of the queue.

View

data ViewF a

View of the front of the queue.

Constructors

a :< !(Steque a)

viewF :: Steque a -> Maybe (ViewF a)

\(\mathcal{O}(1)\) amortized, \(\mathcal{O}(n)\) worst-case. Analyze the front of the queue.

Back

Insert

(|>) :: Steque a -> a -> Steque a infixl 5

\(\mathcal{O}(1)\). Add an element to the back of the queue.

Full queue

Size

null :: Steque a -> Bool

\(\mathcal{O}(1)\). Check if the queue is empty.

size :: Steque a -> Int

\(\mathcal{O}(n)\). Calculate the number of elements stored in the queue. The returned number is guaranteed to be non-negative.

Map

map :: (a -> b) -> Steque a -> Steque b

\(\mathcal{O}(n)\). Apply a function to every element in the queue.

map' :: (a -> b) -> Steque a -> Steque b

\(\mathcal{O}(n)\). Apply a function to every element in the queue and evaluate every resulting element to weak head normal form.

Fold

Left-to-right

foldl :: (b -> a -> b) -> b -> Steque a -> b

\(\mathcal{O}(n_R)\). Fold the queue left-to-right.

foldl' :: (b -> a -> b) -> b -> Steque a -> b

\(\mathcal{O}(n)\). Fold the queue left-to-right with a strict accumulator.

Right-to-left

foldr :: (a -> b -> b) -> b -> Steque a -> b

\(\mathcal{O}(n_L)\). Fold the queue right-to-left.

foldr' :: (a -> b -> b) -> b -> Steque a -> b

\(\mathcal{O}(n)\). Fold the queue right-to-left with a strict accumulator.

Monoid

foldMap :: Monoid m => (a -> m) -> Steque a -> m

\(\mathcal{O}(n_M)\). Map each element of the queue into a monoid, and combine the results.

List

toList :: Steque a -> [a]

\(\mathcal{O}(n)\). Construct a list from a queue. Element order is arbitrary.

toListF :: Steque a -> [a]

\(\mathcal{O}(n)\). Construct a list from a queue. Element order matches queue order, with the head of the list corresponding to the front of the queue.

toListB :: Steque a -> [a]

\(\mathcal{O}(n)\). Construct a list from a queue. Element order matches queue order, with the head of the list corresponding to the back of the queue.

Traverse

traverse :: Applicative f => (a -> f b) -> Steque a -> f (Steque b)

\(\mathcal{O}(n)\). Map each element of a queue to an action, evaluate these actions from front to back, and collect the results.